EditPostForm transfered into PostContainer component for better functionality
